Removing some flash normalization code.
utils.deepReplaceKeyName and utils.translateEventResponse serve a very
specific purpose, which was normalizing output from the flash player before
passing it through to event listeners.

We are experimentally removing deepReplaceKeyName by moving it to embedswf.js
but leaving it commented out.
utils.translateEventResponse has been replaced by a local normalization of floats,
which was it's only practical purpose.

[#88936774]