Rewrite of drag seeking functionality
This change moves the logic for drag seeking (when you click down and slide the cursor along the seekbar)
out of the individual providers and into the javascript controller. This required a few functional changes.
    1. Providers will no longer begin playback after a seek. Instead the controller will decide the correct time to do this.
    2. The "dragging" state is no longer stored in the provider, it is stored in the model.
    3. Since dragging occurs during the "pause" state, we add a visual style flag "jw-flag-dragging" to hide the center play icon.

This fix was required for flash seek dragging to work, but also improves the UX for html due to improvements in how
events are being throttled.

[Fixes #89827172]