Stalled and buffering QOE for media items
There are now two levels of QOE - player level and item level.
The player instance level is a Timer object managed by the API, whereas the item level instance is a Timer object managed by the model. They are called _qoe and _itemQoe respectively.

_itemQoe is reset with each new playlist item, whereas _qoe continues logging data throughout the lifetime of the player instance.

The data from these objects is surfaced via jwplayer().qoe() which returns an object of the form
{
    player : <qoe-data>
    item   : <qoe-data>
}
[Finishes #88512386]