Fix for seeking outside of buffered zone
This requires the provider to perform a seek before the view alerts the model that scrubbing has completed.
The other fix was that the html5 provider shouldn't update state to playing just because we call provider.play(). Instead we first check if it is seeking, and if so, wait until the seek is complete before setting to play.

Included in this fix is a manual stall detection system, given stall events were not being fired by the browser in an expectable manner.

[Fixes #91051588]