Send and track first frame events from html5 player
This extends qoe reporting to include a "first frame" event. This can be triggered by a provider, and if not sent by the provider, it will be automatically triggered as soon as we can tell that the position of the video is increasing.

As part of this change, I separated the QOE reporting into a separate file.

[Finishes #88942902]