Move item title from displayicon to it's own object
We want to move the title and description for a playlist item out of the play button and to the top of the page. This change does so in a new module known as "title" which is owned by the View.

This module watches the model for changes to the item, and when they occur updates its internal text.

[Finishes #90111718]