Store provider state in the mediaModel
In order to reduce confusion over the events being sent out from the model,
we wanted to explicitly separate the state of the player from the state of the
provider.

This is important today, because of the difference between buffering/stalled/loading.
In this expample stalled/loading are important to the model for tracking QOE, whereas
the rest of the player only cares of the player is buffering or not.

After this commit, the mediaModel has an attribute called "state" which represents the
current state of the provider. The Model itself has "state" as well, but it represents
the normalized state.