Separate "complete" state from "idle" state
This change causes providers to set their state to COMPLETE as opposed to
IDLE after playback has concluded.

This distinction is important because the player behaves differently before playback
has began, versus after it has finished. Currently this difference is limited to the
View, however it could become important for plugins related to sharing as well.

[Finishes #92440878]