Allow player to load item of same playlist index
Here I simplify the logic for setting a new playlist item in two ways.
1. Rewrite the formula to use modulo instead of 3 conditional blocks
2. Remove the early return on same playlist index

The second change is important because of a situation where you load a new playlist
while on the first playlist item of a previous playlist. We want to execute this as a
real change, so we will no longer do this check. It is safe to do so, because "repeat"
functionality is handled by the controller now instead of the player.

[Fixes #92550212]