changed the way current and original containers are managed so that it works. update to the unit test, tho the async test may block other unit tests due to its use of video tags (+2 squashed commits) Squashed commits: [9d2c45f] Controller controls the display and removal of views (player view and error screen) and replaces the view content on shutdown. setup errors in embed.js or setup.js sends events which will call the controller.setupError function, which generates and shows the error message views are swapped via controller.showView.  the view is reset by calling showView on the original container
Controller now needs to know the original container when it sets up.  this may cause some issues since we want to maintain the replaced element instead of what may come in when we reset the controller.
[5440279] Makes the view responsible for creating and adding the UI elements of the player to the page instead of making the view and API share the work.
Resetting the view is done via controller.
API defers to getting the view from the controller (which defers to the view) once it has created the view.
completeSetup and replaceOriginalContainer split into different functions
adds aspect ratio via an extra div inside the player instead of a pseudo element.