Implements HD Menu Tooltip
Levels data is put onto the media model for storage, which is what the hd menu listens to for update.
The menu sends toggle
Strategy for setting icons to be active/inactive changed to make them easier to toggle on and off
Tooltips default to hidden.  The can remove their content
Added a player with multiple items in a playlist.  The middle two items have HD options.