Adds Volume Tooltip and Volume Toggling.  Class for the mute icon changed.
Added a tooltip that contains a vertical volume slider.  Implements some code for vertical sliders and adds a toggle to the volume slider icon.  Styles moved out of the slider template file and into the css for the vertical and horizontal slider implementations.
Renamed NewMenu to Menu.
Adds some changed to the classes used for the volume icons.  .jw-icon-volume is the version with sound waves coming out of it.  .jw-icon-volume.jw-off is the icon without waves, indicating muted state.