Adds Captions Menu and changes CC icon class
Adds listeners to create the CC menu when necessary.  also changes the class for the cc icon so that it matches the .jw-off pattern.
This includes a slight hack to get the model to send out the captions list change for the first playlist item since the original event is sent out before the controller is  ready.  will have to discuss the best way to remove this.
Added caption files to the test page.