Handle display icon flickering after seek
This change involved how the seekbar interacts with the player.
When the user clicks down on the time slider, a flag is added to the player
"jw-flag-scrubbing", then the player is paused, and seeks occur until the mouse button
is lifted.
At this point the flag was removed and the player is told to begin playing again.

The bug occurred because the flag was being removed before the player reached a play state.
The solution is to wait until we leave the "pause" state (it will go to either playing or loading)
and then remove the flag.

After testing it, I noticed you could get it to flicker by clicking repeatedly, which is resolved by
not removing the flag in the case that the player has re-entered scrubbing state before the previous
scrubbing event stack had cleared.

[Delivers #93654250]