fix changed urls (+3 squashed commits) Squashed commits: [e0c49a4] Font Color Changes and test updates
* Changed font colors for titles.
* Moved error screen style location out of the error state styles so that it doesn't apply any width or height styling while in the error state
* Added error state to the UI test pages.

[Finishes #92560486]
[920ffd7] Adds a state for errors

Adds an error state to the player that occurs when items can't be loaded or played.  The styles for this state are shared with the startup error state which will occur when the player cannot setup with the information it currently has.

[Finishes #92560486]
[5488a3d] Error screen in progress

Errorscreen changed to a flag so that view can use it to display error information
Error screen mocks the HTML structure and uses the same css to display the error info