Bugfix for calling API methods before setup completes
This was a little tricky because the bug only occurs when setup.setup
completes instantly, triggering the playerReady event without ever deferring
execution. When this happens *and* a plugin or user action uses API methods which
are not in the prototype it would crash.

The deferred execution allowed the controller to fill in the instance methods
which are not in the prototype - such as setControls and addButton.

This fixes the controllers setup to not call setup.setup until these are filled in,
so it can work without relying on the deferred execution of the setup module.

[Fixes #96228308]