Updates to font files
The issue is primarily caused by the font files not being up to date with the most recent version.  icomoon json has the correct svg files in it (play and pause have the same dimensions,volume and volume off icon have the same dimensions, CC on and off have the same dimensions, so on) with consistent spacing around the characters.  SVG files are exported from the project with the correct spacing.

[Finishes #96137652]