Allow clickthrough to provider
The idea is that we want all non-clickable elements above the
provider to ignore pointer-events so that the provider (and ads!)
can receive them.

To make play/pause work when clicking on the video,
we need to pass through click events on the provider (ie when they aren't
handled by ads). These events are passed through to the display, where
they simulate a normal click event.

The other non-intuitive part is that we enable pointer-events when the flag
jw-flag-user-inactive appears. This is so that we don't have to forward
mousemove events from the provider.

[Fixes #96455340]