Remove concept of "showing" from logo
This is a bugfix for the click on a logo not opening up a link.
The problem originated because logo click events always sent a value
of "showing:false" which told the view not to open the link.

However this value was only false because we are no longer toggling
logo visibility programatically, instead using CSS classes.

[Fixes #96118744]