Creates a target for toggling play/pause and allows users to click down to the provider
Renamed displayComp and _display for more accurate
Removing 'visibility: hidden' from setVisibility to allow click events on the provider.  This is also how HTML5 does it.
Maintain listeners on the correct target for touch events
Smarter handling of preventDefault and always bubble events so parent contains can respond to them.
Also fixes issues where you could not drag the screen or scale on mobile when touching the player.

[Finishes #94887836 96677846]