Call provider.remove before instantiating new provider
This issue arose from sharing the video tag between the html5
provider and the shakajs provider. The problem was that when switching
playlist items, the new provider would be constructed before the previous
one was removed from the page.

On construction, the providers look for the presence of a video tag, and if found,
bind events to it. When the previous provider destroys it wipes out the video tag.
Until they learn how to share nicely, all providers will destroy and recreate their
own tag instead of reusing one.

This will need to be revisited for ads in dash playback on mobile,
where you are forced to share a video tag.

[#96552088]