Hide controls a set time after user activity
This unites a few different strategies for hiding the controlbar into one.
Now it is hidden x seconds after the last user activity is recorded.
This fixes the use-case where the mouse would hover over the controlbar and slide
off of the player without hovering over the video element again.

[Fixes #96999912]