Controlbar fades on desktop.  Doesn't fade on mobile.
Removes the display: table; on the jw-flag-ads since it was preventing the controller from naturally being removed when jw-flag-user-inactive was applied.  It was previously there to ensure that the control bar was displayed when we calculate the position of the ad skip button, but that is now the responsibility of the skin.
added a flag for mobile that will make sure that the control bar is displayed during ads.  this flag toggles on and off when its mobile and we're entering instream mode.  _userActivity fired at the beginning of instream mode so that the user can see the new controller state and know its an ad.
!isMobile is added to _componentFadeListeners as we aren't concerned about adding mouse listeners on mobile and it caused a rare issue if the user clicks the control bar a lot during an ad where the control bar wouldn't fade away.

[Finishes #97402648]