Adding the pointer cursor to the video.
Adding the pointer cursor to the video elements and providers.  This implies that the video is clickable for pause/playback toggling.  The pointer is toggled on/off as a result of 'controls' being set to true or false since that also dictates whether you can click to pause.
Flash has its pointer/arrow cursor set based on the controls value.
jw-controls-disabled changed to a flag so it can toggle cursor state.
Moved code that was dictating whether to show the HTML5 native controls to the HTML5 provider.