Place media model in buffering state on instream init
Places the video's mediaModel into the buffering state when instream starts.  This is done to prevent an issue where the media model's state would not change to "playing" when an ad errors out because the old media model was already in the "playing" state.  This also better reflects the actual state of the media at that point  and IDLE (which potentially makes the most sense) causes larger issues.

[Finishes #98862516]