Display icon width and height equal
Sets the display icon width and height to be equal in the core skin (3.5em) and removed padding-based centering.  centering now done via text-align.  Slight bump to the right in the idle mode of the player and slight bump upward on complete mode so that the icon looks a bit more visually centered in its container.

[Finishes #98200440]