Visual centering for play and complete icons
Modified play and complete icons to be more visually centered in the icon.  They're not placed directly in the center of the icon anymore, which could look odd if icons are stacked next to one another without padding, but this is not a use case we have in the player at the moment.