Limited Font Faces in Use, Six doesn't import icons sheet
Reduces the formats of fonts that we're including in the player.  Since IE8 is dropped, we're removing EOT fonts as IE9 supports WOFF.  IE9 in compatibility mode would need EOT fonts, but the player is not functional in compatibility mode there is no point in including them.  WOFF and TTF alone should be sufficient to cover all necessary circumstances.

Removes the @import "../imports/icons" from six since it's causing the broken fonts on the test pages.  It attempts to load fonts which do not exist and the request fails before the browser has finishes decoding the base64 encoded fonts.  This causes the browser to assume the font is broken so it displays character codes instead.  The font afterwords finishes decoding and displays correctly.  Adding the only bit needed from @import "../imports/icons" to the six skin.

[Finishes #99324726]