Fix for volume looping below 0
The intention of the code was to use a default value of 90 when no value was given
for setting volume to the html5 provider. However since it was done with an || instead of
checking if the value is defined, it converted volumes of 0 to 90.

This default value is not necessary since html5Provider.volume is an internal function,
and we can guarantee there will always be a value.

[Fixes #100580948]