JW7-832 JW7-1278 Mobile UI Icon Size Increase
Increases the size of the ui on mobile devices an appropriate amount by adding a flag to the player.  Some extra attention given to the large-ui version of the seven skin so it doesn't look dumb with its extra padding in the small mode.  Also includes a test page with a viewport set to device-width for testing the Mobile UI in a setting more appropriate for the form factor.