JW7-1282 Icon Drawer in Compact Mode
Drawer for extra icons to be placed into when in compact mode.  It toggles when clicking on the more/close icon and can mouse open in small players.   Icons do not currently get taken back out of the drawer if the player gets large enough that it does not need compact mode anymore.  JW7-1283 will complete that task