Do not use 608 captions when webvtt are available
The subtitleTracks event from flash provider is called when the HLS
levels manager detects webvtt captions. When this occurs, we want to
avoid adding in the other types of captions (like 608) which can be defined
in a streams meta data.

This appears to be an aesthetic change, so that the CC list in the menu
is not populated with duplicate or less-desirable caption options.

JW7-1363