Detect when Flash is blocked and remove UI
The goal here is to wait a second after flash has been embedded to hear back from
the flash player. If we don't hear the comfirmation "embedded" event, then we know it's
been blocked. When we enable the plugin it will send the event and player will resume
normally.

JW7-1389