Change state transition for instream item complete
Changing the state to complete as we did before helps us ensure that we get
a play event for each item in a pod of instream items. However it has a side-effect
of causing a replay icon to flicker on screen.

To solve that we only do the "state" change hack when needed. We also use
buffering instead of complete so it will look correct to the user.

JW7-1517