Refactor setup steps
The primary goal of this is to remove model.setItem from setup-steps.js
This will make it easier to remove setItem from the model and move it into the controller.

This change means that you are no longer guaranteed to have a playlistItem when setting up the view,
so there is much more safety around setup methods in the view.