Fix for mediaModel getting reset unexpectedly
This bug manifested as the HD menu not appearing because the controlbar
listens for a "change:levels" event from the mediaModel in order to populate.
It was, however, listening to the wrong instance of mediaModel, because we bound
the listener when we see a "change:playlistItem" event.

We assumed that every time we create a new mediaModel that we would also call
the "change:playlistItem" in order to rebind listeners. This is not true in some
situations, such as when the new item is the same as the old one.

The better solution is to use an event which specifically tells you when the item changes
which I call here "itemSet"

JW7-1563