Playlist loading must happen before providers
This fixes an issue which would arise if an external playlist is loaded, and doesn't
complete loading until after setup steps runs LOAD_YOUTUBE or LOAD_DASH. We actually don't know
if we should use a given provider until after loading the provider and this enforces that.

This also removes setPlaylist from the public API, preferring to pass a reference to the method.