Fix for HD menu not showing up
This bug manifested because the mediaModel was getting changed without
the controlbar updating it's mediaModel event listeners. This happens
in a situation where we set a new playlist item to be the same as
the old one.

In that situation there is no "change:playlistItem" event, even though
the mediaModel gets reset.

JW7-1563