Allow player to resume after detach during seek
When we switch to instream mode while the player is mid-seek
it was causing the state to get stuck in the "seeking" mode, which prevented
the provider from resuming when instream mode ended. Instead, it would wait to
play for the "seeked" event to occur, which it never does.

JW7-1665