Fix for changing default caption label
We saw an error where new defaults were not working when
changing the item but keeping the same provider. This is a quirk
of how providers use the players config by cloning an instance
of the model when the provider is instantiated.

This quirk would be eliminated by passing in a getter method
to providers which gets the current state of the provider, instead
of passing in a simple object.

JW7-1719