Only do the pointer-events hack for flash provider
We currently have a work-around which disables and re-enables
pointerEvents on the play button on mousedown and mouseup.
This allows a click to touch a swf which is behind the play button
allowing chrome to know it was a true click and not throttle playback.

Since this work-around is only for Chrome-Flash there is no reason to
use it for Chrome-non-flash.

JW7-1759