Change oldProvider state when instream destorys
When instream is created, we change model's mediaModel state.
When the ad errors out, the instream destroys and tries to start the video.
Since the mediaModel's state is buffering, it changes the state to playing, but the provider's state is still playing.
This causes the mediaModel's state to stay in buffering, causing buffer icon to display on mobile devices.
JW7-1739