Try/Catch on api.trigger instead of api.on
Our first approach to safely handling user callbacks was to wrap
the users code in a try/catch, and calling that our new safeCallback.
This solution does not work because when the user tries api.off('event', fn)
the fn no longer matches the function in api._events and so the off command doesn't
remove the bound event.

This fixes that by try/catching around when the callback is called in the trigger method.
A neat side-effect of this is that we don't need to overwrite the .on method at all, so
I went ahead and cleaned out some older debug code as well.

JW7-1764