Compact UI maintains compact state more often
State updates changed to allow a 20% width control bar.
clearCompactMode will recheck whether compact mode should be maintained, which will automatically update compact mode when related changes are made.  This will make pop-in of a different style control bar after user input or ads not happen as often
Make logic to switch states not guarded by the current state.  This will make it choose a state and stick to it better.  Some of the comparisons changed as well.

JW7-1662