Fix playlist error handling after api.load
This bug manifested when calling api.load on a playlist of which all sources get
filtered out. For example, an rtmp file on an android chrome.

In this case, the player was not going to an error state, instead it was sending
an error through the mediaController. The solution was to send the error through
the actual model.

JW7-1783