Supprt JSON formatted playlist
When playlist loader sends ajax request, ajax used to send out error if it cannot parse the response xml.
Instead, we want to complete the ajax request if there is a responseText.
Then, when rss parsing fails in playlist loader, we try to use the responseText to parse that into JSON.
Added a unit test for JSON feed playlist.
JW7-1206