Fix 'Live Broadcast' issue with item change
Since we get timeupdate event with playlist item change, we update the duration for androidHLS on item change.
For hls, we still have duration as infinity, resulting 'Live Broadcast' message bug.
Making a fix so that if the videotag's currentTime is 0(not yet playing), we do not update the duration.
JW7-1789