Fix RTMP streams starting over after midrolls
In RTMP play method, we resume the video if the video is paused, but never set the state to playing.
Controller's play function is called a few times after a midroll, and checks for the state before calling provider's play function.
Since the state is not updated, controller called the play function, resulting the stream to start playing again from the beginning.

JW7-1877