Check for, and clear tracks on loaded data event
This puts the event handling and check for track presence in _setAudioTracks and _setTextTracks. The goal is to keep _onLoadedData clean and abstract all the work away into those setters.

It's important that we listen for changes for triggering events. Triggering of any event should always be as for down the stack as possible so that any actions taken by event listeners do not impact code further down the stack or escape unregistered callbacks.