Stop encoding urls for preview images
This was originally done to allow a publisher who passed in non-encoded urls
to still work (specifically, they were passing a url which has a double-quote character).

This causes a problem for other users who have already encoded the url, since double-encoding
breaks pathing. A better solution is to require users to pass in a well-formed url.

https://github.com/jwplayer/jwplayer/issues/919
JW7-883