Fix flash reload same item not working
In load function for videoMediaProvider, the function did not do anything if the item was the same and preload is not equal to auto.
This resulted the provider to be stuck on buffering when reloading the same item.
Adding an else statement to prevent the provider from being stuck.

JW7-1821