Support HLS ID3 metadata in Edge
Since at least EdgeHTML 13.10586, Edge reads ID3 metadata from HLS streams, but the format exposed to JavaScript is different from that which Safari uses. This commit adds code to detect Edge's format and convert it to match Safari's before doing further processing.

The "if (array[i] >= 32)" statement is there because for some unknown reason, the data field of my cue points consistently starts with a 0x03 character.

Edge will also sometimes expose more than one active cue point to JavaScript. This commit does not address that, so if two cue points are active with the same data field, JW Player may not use the more recent of the two.