Detect when fonts are loaded before displaying icons
Detect when fonts are loaded before displaying icons by polling the width of a span.  The span starts out with a basic width and will have its size increase ones the icon font is applied and wider glyphs are used instead of the default glyphs.  This may be a more appropriate way to determine if the UI is ready in the future to make sure we don't need to do many repaints before we display the UI.

JW7-2481