Handle complete event outside on all
When an event is triggered, on all listener is called after the on event listener.
This caused the instreamItemComplete function to be called before on all and update the instream index.
The on all listener then forwards the complete event, having the complete event to have a wrong instream index.
Handling to complete event with instreamItemComplete so that we can control over this behavior.
JW7-2296