Add width and height to style when resizing
Some VPAID ad resize with an incorrect width.
This is fine, but when we resize again after the VPAID ad, since there is no width/height associated with the style,
we resize with a wrong width/height.
Adding video width and height to the style as empty to begin with, so that we reset the width and height.
JW7-2625