Add filter to sources with the chosen provider
When the stream type is hls, but the source has fairplay drm, we need to filter the source.
However, we used to only filter the source by bestType.
This resulted the first source with hls fairplay to be not filtered in caterpillar provider if the second source is hls,
since it chose hls as the bestType.
Adding a second filter to filterSource function so that we filter sources by the provider again when bestType is chosen.
JW7-2841