Adjust control bar height for live broadcast
By default the control bar height is set to accommodate live broadcast
and the override is for anything that isn’t live broadcast. The delay
the player has before it detects a live broadcast and adds causes a
flicker. Handling control bar in this way fixes that experience in time
slider above mode.