Display alt text correctly at largest and smallest breakpoints
Ensures alt text is not cut of in live mode by setting the touch target width of content only if it is not the .jw-alt-text container.  This allows the alt text container to size correctly without needing to override styles.  Also fixes issues with the duration value appearing during live mode in smaller display sizes in live streams.

JW7-3705