Add :not()s for plugins to keep affecting styles in one place
Add :not()s for plugins which are affected by this change (related and sharing) to keep styles relating to time-slider-above changes in a single location instead of forcing specificity in the css in their repos.  This should keep things easier to follow and prevent issues where too many overrides are added.

JW7-3666