Fit jw-icon-rewind:before to the size of the icon for correct safari highlights and display
Sets the 10 second rewind button's before pseudo element to fill its associated element.  This makes it so Safari will allow the entire icon can highlight correctly because Safari will only set the color on the part of the glyph that exists in that container.  This also fixes the issue where part of the icon is cut off in the audio only mode because it does not have a large enough line-height set.

JW7-3912 JW7-3664