Add conditionals to color overrides
Conditionally applies colorization values if they are set.  This makes it so that "background: none" will not be applied by the shorthand property if we do not have an active, inactive or background color defined.  Previously, no value would exist when setting background color and would then be ignored, but with the default addition of "none" we need to make sure these are only applied conditionally.
Removes the if(!value) check as it is no longer used.

JW7-3908