Add thumbstrip functionality to touch and PointerEvent devices
Adds thumbstrip functionality to touch and PointerEvent devices by activating listeners for touch and mouse events instead of just mouse events.  In order to support touch platforms which will not activate the mouseover handler that will set the this.activeCue property, we have to use javascript to check what the closest cue is in the hover/drag handler.  We will set a cue that is within 5 pixels away to be the active cue so that we can show the text for the cue if necessary.

JW7-3167