ESLint underscore.js
I just happened to come across a duplicate decelration of `_.any` in
`underscore.js`, so decided to lint the file. I only temporarily
removed it from the ignore list to lint, because I did not want to
change any of logic (such as using `===` instead of `==`), and only
made the stylist changes (mostly adding curly braces around `if`
statements.