Add option to not pause when instream init
Googima iOS pauses the ad when the video tag is paused on ad init,
because the same video tag is used to play the ad and the content.
Add a flag to not pause the video tag on instream init.
The same code was there but reverted after a fix, but a change from IMA SDK may have caused the problem to occur again.
The commit that reverted the change: https://github.com/jwplayer/jwplayer/commit/786dce1831810f24557831855eea04bfd86ca881
ADS-187