Remove setPlaybackRate`_queueCommand` from controller
`setPlaybackRate` should not be queued in the controller. Queued commands require a method on the controller that starts with underscore that will be called when a provider is available. Since this method sets a model property, it does not require a provider to be instantiated for the call to work, and thus does not need to be queued.