Fix "playReason" when autostart is on (#2245)
Add a cancelable module used to handle checking autostart after the event loop is complete.

Moves setting of "playAttempt" to true on the model until after "beforePlay" is triggered. Setting "playAttempt" to true triggers the start of playback if the current provider has fired "bufferFull".

Clears the model's "playlistItem" attribute before setting it to ensure a change event is fired from the model, triggering "playlistItem" from the player api. (We used to fire this on "itemReady", but playlistItem should fire as soon as "playlistItem" is set and media model properties are reset).


JW8-313
JW8-312
JW8-309