Use model.setProvider promise when calling play() and provider is not defined (#2450)
This PR will...

Use model.setProvider promise when calling play() and provider is not defined.

Why is this Pull Request needed?

There is a different promise chain for setting a playlist item and for playing the current item. Because setting an item will start loading a provider when necessary, we don't need to request that a provider be loaded when play() is requested. We should use the promise created when the item was set.

Addresses Issue(s):

JW8-663